revert CFG move
