+++ ultralytics/yolo/cfg/default.yaml
